refactor: move docsite examples to examples package
