+++ packages/docs/examples/src/renderless/AreaChart.tsx
@@ -28,7 +28,7 @@ const data = [
 /**
  * Adapted from https://vega.github.io/vega/examples/bar-chart/
  */
-const AreaChart: React.FC = memo(() => (
+export const AreaChart: React.FC = memo(() => (
 	<Chart width={500} height={200} padding={8} data={{ data }}>
 		<LinearScale
 			name="x"
@@ -52,4 +52,3 @@ const AreaChart: React.FC = memo(() => (
 	</Chart>
 ))
 AreaChart.displayName = 'AreaChart'
-export default AreaChart
