+++ packages/docs/examples/src/renderless/index.ts
@@ -1,4 +1,9 @@
+export * from './AreaChart'
 export * from './BarChart'
+export * from './BarChartWithMean'
+export * from './ElbowPlot'
 export * from './GroupedBarChart'
+export * from './JobVoyager'
 export * from './LineChart'
+export * from './StackedAreaChart'
 export * from './StackedBarChart'
