+++ packages/docs/examplesweb/stories/renderless-examples/index.tsx
@@ -6,9 +6,14 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 import {
+	Areachart,
 	BarChart,
+	BarChartWithMean,
+	ElbowPlot,
 	GroupedBarChart,
+	JobVoyager,
 	LineChart,
+	StackedAreaChart,
 	StackedBarChart,
 } from '@chart-parts/examples'
 import { Renderer } from '@chart-parts/react-svg-renderer'
@@ -17,14 +22,24 @@ import { ChartingProvider } from '@chart-parts/react'
 const renderer = new Renderer()
 
 storiesOf('Renderless React Examples', module)
+	.add('Area Chart', () => (
+		<ChartingProvider value={renderer}>
+			<AreaChart />
+		</ChartingProvider>
+	))
 	.add('Bar Chart', () => (
 		<ChartingProvider value={renderer}>
 			<BarChart />
 		</ChartingProvider>
 	))
-	.add('Stacked Bar Chart', () => (
+	.add('Bar Chart with Mean', () => (
 		<ChartingProvider value={renderer}>
-			<StackedBarChart />
+			<BarChartWithMean />
+		</ChartingProvider>
+	))
+	.add('Elbow Plot', () => (
+		<ChartingProvider value={renderer}>
+			<ElbowPlot />
 		</ChartingProvider>
 	))
 	.add('Grouped Bar Chart', () => (
@@ -32,8 +47,23 @@ storiesOf('Renderless React Examples', module)
 			<GroupedBarChart />
 		</ChartingProvider>
 	))
+	.add('Job Voyager', () => (
+		<ChartingProvider value={renderer}>
+			<JobVoyager />
+		</ChartingProvider>
+	))
 	.add('Line Chart', () => (
 		<ChartingProvider value={renderer}>
 			<LineChart />
+		</ChartingProvider>
+	))
+	.add('Stacked Area Chart ', () => (
+		<ChartingProvider value={renderer}>
+			<StackedAreaChart />
+		</ChartingProvider>
+	))
+	.add('Stacked Bar Chart', () => (
+		<ChartingProvider value={renderer}>
+			<StackedBarChart />
 		</ChartingProvider>
 	))
